// src/generated/models/Cr38e_projectsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr38e_projectscr38e_phase = {
  357090000: 'Planning',
  357090001: 'Design',
  357090002: 'Development',
  357090003: 'Testing',
  357090004: 'Deployment',
  357090005: 'Closed'
} as const;
export type Cr38e_projectscr38e_phase = keyof typeof Cr38e_projectscr38e_phase;
export const Cr38e_projectscr38e_projecttype = {
  357090000: 'Internal',
  357090001: 'External',
  357090002: 'Capital',
  357090003: 'Operational'
} as const;
export type Cr38e_projectscr38e_projecttype = keyof typeof Cr38e_projectscr38e_projecttype;
export const Cr38e_projectsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr38e_projectsstatecode = keyof typeof Cr38e_projectsstatecode;
export const Cr38e_projectsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr38e_projectsstatuscode = keyof typeof Cr38e_projectsstatuscode;

export interface Cr38e_projectsBase {
  cr38e_datacontrol?: string;
  cr38e_description?: string;
  cr38e_phase: Cr38e_projectscr38e_phase;
  cr38e_projectid: string;
  cr38e_projectname: string;
  cr38e_projectscope?: string;
  cr38e_projecttype: Cr38e_projectscr38e_projecttype;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr38e_projectsstatecode;
  statuscode?: Cr38e_projectsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr38e_projects extends Cr38e_projectsBase {
  cr38e_phasename?: string;
  cr38e_projecttypename?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}